+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.6.1-dev
  - replace --print-stats with --no-stats, default to printing stats
+ - make gaggle an optional compile-time feature
 
 ## 0.6.0 May 14, 2020
  - Initial support for gaggles: distributed load testing