+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## 0.7.0-dev
+ - async
+
 ## 0.6.1 May 16, 2020
  - replace --print-stats with --no-stats, default to printing stats
  - make gaggle an optional compile-time feature